update: chat에 라우팅페이지 삭제한후에, 쿼리로 다이나믹페이지 만듬, DM페이지도 비슷하게 할예정
