+++ front/pages/Chat.tsx
@@ -10,25 +10,83 @@ import useSWR from "swr";
 import axios from "axios";
 import { GetServerSideProps } from "next";
 import useSocket from "../component/Utils/socket";
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
+import fetcher from "../component/Utils/fetcher";
+import PWModal from "../component/ChatRoom/PWModal";
+import ChatRoomBody from "../component/ChatRoom/ChatRoomBody";
 
-export default function Chat({ accessToken }: { accessToken: string }) {
-  const { data, error } = useSWR("/api/users");
+export default function Chat({
+  accessToken,
+  id,
+}: {
+  accessToken: string;
+  id: any;
+}) {
+  const isId = Object.keys(id).length !== 0;
+  const { data: userData, error: userError } = useSWR("/api/users");
+  // roomData를 가지고, 오너가 볼수있는 모달을 추가하고, 비밀번호 추가하고
+  const { data: roomData, error: roomError } = useSWR(
+    isId ? `/api/chatroom/${id.id}` : null,
+    isId ? fetcher : null
+  );
   const [socket] = useSocket(accessToken, "chat");
+  const [showPWModal, setShowPWModal] = useState<boolean>(true);
+  console.log(id);
+  // 룸 데이터 이동 확인용, 모달용
+  useEffect(() => {
+    if (roomData) {
+      console.log(`i am in room ${roomData.chatroomName}`);
+    }
+    return () => {
+      if (roomData) {
+        console.log(`1 am out room ${roomData.chatroomName}`);
+      }
+      setShowPWModal(true);
+    };
+  }, [roomData, isId]);
 
-  if (error) axios.get("/api/auth/refresh").catch((e) => console.log(e));
-  if (!data || !socket) return <Loading />;
+  if (userError || (id.id && roomError))
+    axios.get("/api/auth/refresh").catch((e) => console.log(e));
+  if (!userData || (id.id && !roomData) || !socket) return <Loading />;
   return (
     <Layout>
-      <Title title="Chat" />
-      {data.two_factor_activated && !data.two_factor_valid && (
+      <Title title="ChatRoom" />
+      {userData.two_factor_activated && !userData.two_factor_valid && (
         <TwoFactorModal />
       )}
-      <div style={{ display: "grid", gridTemplateColumns: "2fr 4fr 2fr" }}>
+      {isId &&
+        roomData &&
+        roomData.password &&
+        showPWModal &&
+        roomData.ownerId !== userData.id && (
+          <PWModal
+            setShowPWModal={setShowPWModal}
+            password={roomData.password}
+          />
+        )}
+      <div className="component-style">
         <RoomList accessToken={accessToken} />
-        <ChatBody />
+        {/* 일반 화면 */}
+        {!isId && <ChatBody />}
+        {/* 채팅룸 */}
+        {isId && id.link === "chat" && <ChatRoomBody chatroomId={id.id} />}
         <Participant />
       </div>
+      <style jsx>{`
+        .pwmodal-background {
+          position: fixed;
+          top: 0;
+          left: 0;
+          bottom: 0;
+          right: 0;
+          background: rgba(0, 0, 0, 0.8);
+          z-index: 1;
+        }
+        .component-style {
+          display: grid;
+          grid-template-columns: 2fr 4fr 2fr;
+        }
+      `}</style>
     </Layout>
   );
 }
@@ -36,6 +94,7 @@ export default function Chat({ accessToken }: { accessToken: string }) {
 export const getServerSideProps: GetServerSideProps = async (context) => {
   const cookie = cookies(context);
   const { accessToken } = cookie;
+  const id = context.query;
   if (!accessToken) {
     return {
       redirect: {
@@ -45,5 +104,5 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
     };
   }
   // tokenManager(cookie);
-  return { props: { accessToken } };
+  return { props: { accessToken, id } };
 };
